Trim fclShopMisc FUN_003ff460 parameter width

+++ src/Event/Fcl/fclShopMisc.c
@@ -377,7 +377,7 @@ u64 FUN_003fee40(int param_1);
 u64 FUN_003ff150_i(int param_1,int param_2);
 u32 FUN_003fef30(int param_1,int param_2);
 u64 FUN_003ff150(u64 param_1,long param_2);
-void FUN_003ff460(int param_1,u64 param_2);
+void FUN_003ff460(int param_1,u32 param_2);
 u32 FUN_003ff570(u64 param_1);
 void FUN_003ff630(u32 param_1,u64 param_2,u64 param_3,u64 param_4, u8 param_5,short param_6,short param_7,u32 param_8);
 u8 FUN_003ffa20(u64 param_1,u64 param_2,int param_3);
@@ -12101,7 +12101,7 @@ u64 FUN_003ff150(u64 param_1,long param_2)
 // FUN_003FF460 NONMATCHING
 
 
-void FUN_003ff460(int param_1,u64 param_2)
+void FUN_003ff460(int param_1,u32 param_2)
